requirements.txt
pillow
numpy
cython
six
git+https://github.com/HarvestDigitalE/pytoshop.git

